Refactor collection removal logic and event handling
- CadObjectCollection<T>.Remove now returns bool and supports OnBeforeRemove for cancellation
- CollectionChangedEventArgs gains Cancel property for event-driven operation control
- Prevent removal of paper space viewport via BlockRecord event
- Update Viewport.RepresentsPaper logic for accuracy
- Refactor tests and IO code to use new Remove signature
- Restore DxfDocumentBuilder and update SeqendCollection<T> for new patterns
- Use AssociatedBlock in Layout for clarity and consistency
