+++ src/ACadSharp/CadObjectCollection.cs
@@ -4,123 +4,144 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace ACadSharp
+namespace ACadSharp;
+
+/// <summary>
+/// Collection formed by <see cref="CadObject"/> managed by an owner.
+/// </summary>
+/// <typeparam name="T"></typeparam>
+public class CadObjectCollection<T> : IObservableCadCollection<T>
+	where T : CadObject
 {
+	/// <inheritdoc/>
+	public event EventHandler<CollectionChangedEventArgs> OnAdd;
+
 	/// <summary>
-	/// Collection formed by <see cref="CadObject"/> managed by an owner.
+	/// Occurs before an item is removed from the collection.
 	/// </summary>
-	/// <typeparam name="T"></typeparam>
-	public class CadObjectCollection<T> : IObservableCadCollection<T>
-		where T : CadObject
+	/// <remarks>Subscribers can use this event to perform actions or validation before the removal operation
+	/// completes. To cancel the removal, handle the event and set the appropriate property on the event arguments if
+	/// supported.</remarks>
+	public event EventHandler<CollectionChangedEventArgs> OnBeforeRemove;
+
+	/// <inheritdoc/>
+	public event EventHandler<CollectionChangedEventArgs> OnRemove;
+
+	/// <summary>
+	/// Gets the number of elements that are contained in the collection.
+	/// </summary>
+	public int Count { get { return this._entries.Count; } }
+
+	/// <summary>
+	/// Owner of the collection.
+	/// </summary>
+	public CadObject Owner { get; }
+
+	protected readonly HashSet<T> _entries = new HashSet<T>();
+
+	/// <summary>
+	/// Default constructor for a <see cref="CadObjectCollection{T}"/> with it's owner assigned.
+	/// </summary>
+	/// <param name="owner">Owner of the collection.</param>
+	public CadObjectCollection(CadObject owner)
 	{
-		/// <inheritdoc/>
-		public event EventHandler<CollectionChangedEventArgs> OnAdd;
+		this.Owner = owner;
+	}
 
-		/// <inheritdoc/>
-		public event EventHandler<CollectionChangedEventArgs> OnRemove;
+	/// <summary>
+	/// Add a <see cref="CadObject"/> to the collection, this method triggers <see cref="OnAdd"/>.
+	/// </summary>
+	/// <param name="item"></param>
+	/// <exception cref="ArgumentException"></exception>
+	/// <exception cref="ArgumentNullException"></exception>
+	public virtual void Add(T item)
+	{
+		if (item is null) throw new ArgumentNullException(nameof(item));
 
-		/// <summary>
-		/// Owner of the collection.
-		/// </summary>
-		public CadObject Owner { get; }
+		if (item.Owner != null)
+			throw new ArgumentException($"Item {item.GetType().FullName} already has an owner", nameof(item));
 
-		/// <summary>
-		/// Gets the number of elements that are contained in the collection.
-		/// </summary>
-		public int Count { get { return this._entries.Count; } }
+		if (this._entries.Contains(item))
+			throw new ArgumentException($"Item {item.GetType().FullName} is already in the collection", nameof(item));
 
-		public T this[int index]
+		this._entries.Add(item);
+		item.Owner = this.Owner;
+
+		OnAdd?.Invoke(this, new CollectionChangedEventArgs(item));
+	}
+
+	/// <summary>
+	/// Add multiple <see cref="CadObject"/> to the collection, this method triggers <see cref="OnAdd"/>.
+	/// </summary>
+	/// <param name="items"></param>
+	public void AddRange(IEnumerable<T> items)
+	{
+		foreach (var item in items)
 		{
-			get
+			this.Add(item);
+		}
+	}
+
+	/// <summary>
+	/// Removes all elements from the Collection.
+	/// </summary>
+	public void Clear()
+	{
+		Queue<T> q = new(this._entries.ToList());
+		while (q.TryDequeue(out T entry))
+		{
+			this.Remove(entry);
+		}
+	}
+
+	/// <inheritdoc/>
+	public IEnumerator<T> GetEnumerator()
+	{
+		return this._entries.GetEnumerator();
+	}
+
+	/// <inheritdoc/>
+	IEnumerator IEnumerable.GetEnumerator()
+	{
+		return this._entries.GetEnumerator();
+	}
+
+	/// <summary>
+	/// Removes the specified item from the collection.
+	/// </summary>
+	/// <remarks>If an event handler for the removal is registered and cancels the operation, the item will not be
+	/// removed. After successful removal, the item's owner is set to null and a removal event is raised.</remarks>
+	/// <param name="item">The item to remove from the collection.</param>
+	/// <returns>true if the item was successfully removed; otherwise, false.</returns>
+	public virtual bool Remove(T item)
+	{
+		if (this.OnBeforeRemove != null)
+		{
+			CollectionChangedEventArgs args = new(item);
+			this.OnBeforeRemove.Invoke(this, args);
+			if (args.Cancel)
 			{
-				return this._entries.ElementAtOrDefault(index);
+				return false;
 			}
 		}
 
-		protected readonly HashSet<T> _entries = new HashSet<T>();
-
-		/// <summary>
-		/// Default constructor for a <see cref="CadObjectCollection{T}"/> with it's owner assigned.
-		/// </summary>
-		/// <param name="owner">Owner of the collection.</param>
-		public CadObjectCollection(CadObject owner)
+		if (!this._entries.Remove(item))
 		{
-			this.Owner = owner;
+			return false;
 		}
 
-		/// <summary>
-		/// Add a <see cref="CadObject"/> to the collection, this method triggers <see cref="OnAdd"/>.
-		/// </summary>
-		/// <param name="item"></param>
-		/// <exception cref="ArgumentException"></exception>
-		/// <exception cref="ArgumentNullException"></exception>
-		public virtual void Add(T item)
+		item.Owner = null;
+
+		OnRemove?.Invoke(this, new CollectionChangedEventArgs(item));
+
+		return true;
+	}
+
+	public T this[int index]
+	{
+		get
 		{
-			if (item is null) throw new ArgumentNullException(nameof(item));
-
-			if (item.Owner != null)
-				throw new ArgumentException($"Item {item.GetType().FullName} already has an owner", nameof(item));
-
-			if (this._entries.Contains(item))
-				throw new ArgumentException($"Item {item.GetType().FullName} is already in the collection", nameof(item));
-
-			this._entries.Add(item);
-			item.Owner = this.Owner;
-
-			OnAdd?.Invoke(this, new CollectionChangedEventArgs(item));
-		}
-
-		/// <summary>
-		/// Add multiple <see cref="CadObject"/> to the collection, this method triggers <see cref="OnAdd"/>.
-		/// </summary>
-		/// <param name="items"></param>
-		public void AddRange(IEnumerable<T> items)
-		{
-			foreach (var item in items)
-			{
-				this.Add(item);
-			}
-		}
-
-		/// <summary>
-		/// Removes all elements from the Collection.
-		/// </summary>
-		public void Clear()
-		{
-			Queue<T> q = new(this._entries.ToList());
-			while (q.TryDequeue(out T entry))
-			{
-				this.Remove(entry);
-			}
-		}
-
-		/// <summary>
-		/// Removes a <see cref="CadObject"/> from the collection, this method triggers <see cref="OnRemove"/>.
-		/// </summary>
-		/// <param name="item"></param>
-		/// <returns>The removed <see cref="CadObject"/></returns>
-		public virtual T Remove(T item)
-		{
-			if (!this._entries.Remove(item))
-				return null;
-
-			item.Owner = null;
-
-			OnRemove?.Invoke(this, new CollectionChangedEventArgs(item));
-
-			return item;
-		}
-
-		/// <inheritdoc/>
-		public IEnumerator<T> GetEnumerator()
-		{
-			return this._entries.GetEnumerator();
-		}
-
-		/// <inheritdoc/>
-		IEnumerator IEnumerable.GetEnumerator()
-		{
-			return this._entries.GetEnumerator();
+			return this._entries.ElementAtOrDefault(index);
 		}
 	}
 }
