+++ src/ACadSharp/IObservableCadCollection.cs
@@ -1,23 +1,22 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace ACadSharp
+namespace ACadSharp;
+
+/// <summary>
+/// Events for an observed cad collection.
+/// </summary>
+/// <typeparam name="T"></typeparam>
+public interface IObservableCadCollection<T> : IEnumerable<T>
+	where T : CadObject
 {
 	/// <summary>
-	/// Events for an observed cad collection.
+	/// Event triggers when an object is added to the collection.
 	/// </summary>
-	/// <typeparam name="T"></typeparam>
-	public interface IObservableCadCollection<T> : IEnumerable<T>
-		where T : CadObject
-	{
-		/// <summary>
-		/// Event triggerrs when an object is added to the collection.
-		/// </summary>
-		event EventHandler<CollectionChangedEventArgs> OnAdd;
+	event EventHandler<CollectionChangedEventArgs> OnAdd;
 
-		/// <summary>
-		/// Event triggerrs when an object is removed from the collection.
-		/// </summary>
-		event EventHandler<CollectionChangedEventArgs> OnRemove;
-	}
+	/// <summary>
+	/// Event triggers when an object is removed from the collection.
+	/// </summary>
+	event EventHandler<CollectionChangedEventArgs> OnRemove;
 }
