+++ src/ACadSharp/SeqendCollection.cs
@@ -2,73 +2,71 @@
 using System;
 using System.Linq;
 
-namespace ACadSharp
+namespace ACadSharp;
+
+/// <summary>
+/// Represents a collection of <see cref="CadObject"/> ended by a <see cref="Entities.Seqend"/> entity.
+/// </summary>
+/// <typeparam name="T"></typeparam>
+public class SeqendCollection<T> : CadObjectCollection<T>, ISeqendCollection
+	where T : CadObject
 {
+	public event EventHandler<CollectionChangedEventArgs> OnSeqendAdded;
+
+	public event EventHandler<CollectionChangedEventArgs> OnSeqendRemoved;
+
 	/// <summary>
-	/// Represents a collection of <see cref="CadObject"/> ended by a <see cref="Entities.Seqend"/> entity.
+	/// Sequence end entity for dxf.
 	/// </summary>
-	/// <typeparam name="T"></typeparam>
-	public class SeqendCollection<T> : CadObjectCollection<T>, ISeqendCollection
-		where T : CadObject
+	public Seqend Seqend
 	{
-		public event EventHandler<CollectionChangedEventArgs> OnSeqendAdded;
-
-		public event EventHandler<CollectionChangedEventArgs> OnSeqendRemoved;
-
-		/// <summary>
-		/// Sequence end entity for dxf.
-		/// </summary>
-		public Seqend Seqend
+		get
 		{
-			get
-			{
-				if (this._entries.Any())
-					return this._seqend;
-				else
-					return null;
-			}
-			internal set
-			{
-				this._seqend = value;
-				this._seqend.Owner = this.Owner;
-			}
+			if (this._entries.Any())
+				return this._seqend;
+			else
+				return null;
+		}
+		internal set
+		{
+			this._seqend = value;
+			this._seqend.Owner = this.Owner;
+		}
+	}
+
+	private Seqend _seqend;
+
+	public SeqendCollection(CadObject owner) : base(owner)
+	{
+		this._seqend = new Seqend(owner);
+	}
+
+	/// <inheritdoc/>
+	public override void Add(T item)
+	{
+		bool addSeqend = false;
+		if (!this._entries.Any())
+		{
+			addSeqend = true;
 		}
 
-		private Seqend _seqend;
+		base.Add(item);
 
-		public SeqendCollection(CadObject owner) : base(owner)
+		// The add could fail due an Exception
+		if (addSeqend && this._entries.Any())
 		{
-			this._seqend = new Seqend(owner);
+			this.OnSeqendAdded?.Invoke(this, new CollectionChangedEventArgs(this._seqend));
+		}
+	}
+
+	/// <inheritdoc/>
+	public override bool Remove(T item)
+	{
+		if(base.Remove(item))
+		{
+			this.OnSeqendRemoved?.Invoke(this, new CollectionChangedEventArgs(this._seqend));
 		}
 
-		/// <inheritdoc/>
-		public override void Add(T item)
-		{
-			bool addSeqend = false;
-			if (!this._entries.Any())
-			{
-				addSeqend = true;
-			}
-
-			base.Add(item);
-
-			// The add could fail due an Exception
-			if (addSeqend && this._entries.Any())
-			{
-				this.OnSeqendAdded?.Invoke(this, new CollectionChangedEventArgs(this._seqend));
-			}
-		}
-
-		/// <inheritdoc/>
-		public override T Remove(T item)
-		{
-			var e = base.Remove(item);
-			if(e != null)
-			{
-				this.OnSeqendRemoved?.Invoke(this, new CollectionChangedEventArgs(this._seqend));
-			}
-
-			return e;
-		}
+		return true;
 	}
 }
